moved prf related functions to its module
